CI: fix build errors due to numpy and dashu 06/21 releases (#2772)
The mypy error is due to numpy. It does not fail on python 3.10 because
it does not update to 2.5.0 for that python version. For python 3.13
(mbi) because mbi brings in jax and numpy.  numpy updates to 2.5.0 and
causes the mypy error. mypy stays at an older pin that does not support
the new numpy typing.

The dashu fix is a direct pin to avoid the new 04.5 version for dashu
float*.

The mypy fix is a bit more involved. Rather than splitting requirements
across both python version(can work but has drawbacks), I update the .in
compliing to make a universal requirements.txt. This allows the file to
handle different python versions accordingly. It does come at the cost
of requiring uv for compilation pip-compiling does not support universal
yet,(according to some preliminary checking I did).

I removed 2 files that seemed unused and were giving mypy errors while
debugging. Those can be reverted but seemed like artifacts of kicking
off the project

---------

+++ python/test/mbi/test_table.py
@@ -54,7 +54,7 @@ def test_fit_effectiveness(algorithm, privacy_loss, approximate):
     data_lf = pl.LazyFrame(data_dict)
     lookup_B = {"a": -2.5, "b": -1.5, "c": -0.5, "d": 0.5, "e": 1.5, "f": 2.5, None: 0}
     lookup_C = dict(zip(range(8), np.linspace(-3.5, 3.5, 8)))
-    lookup_C[None] = 0  # type: ignore[index]
+    lookup_C[None] = np.float64(0)  # type: ignore[index]
 
     roundtrip_np = np.stack(
         [
